Allow pyroscope agent to run without a config file
