+++ pkg/cli/agent_windows.go
@@ -2,21 +2,16 @@ package cli
 
 import (
 	"fmt"
+	"io/ioutil"
+	"os"
 
 	"github.com/kardianos/service"
 	"github.com/sirupsen/logrus"
+	"gopkg.in/yaml.v2"
 
 	"github.com/pyroscope-io/pyroscope/pkg/agent/target"
 	"github.com/pyroscope-io/pyroscope/pkg/config"
 )
-
-func startAgent(c *config.Agent) error {
-	if err := loadTargets(c); err != nil {
-		return fmt.Errorf("configuration is invalid: %w", err)
-	}
-	svc, _ := service.New(newAgentService(c), &service.Config{Name: "pyroscope"})
-	return svc.Run()
-}
 
 type agentService struct{ tgtMgr *target.Manager }
 
@@ -34,3 +29,28 @@ func (svc agentService) Stop(_ service.Service) error {
 }
 
 func (svc agentService) start() error { return svc.tgtMgr.Start() }
+
+func startAgent(c *config.Agent) error {
+	if err := loadTargets(c); err != nil {
+		return fmt.Errorf("configuration is invalid: %w", err)
+	}
+	svc, _ := service.New(newAgentService(c), &service.Config{Name: "pyroscope"})
+	return svc.Run()
+}
+
+func loadTargets(c *config.Agent) error {
+	b, err := ioutil.ReadFile(c.Config)
+	switch {
+	case err == nil:
+	case os.IsNotExist(err):
+		return nil
+	default:
+		return err
+	}
+	var a config.Agent
+	if err = yaml.Unmarshal(b, &a); err != nil {
+		return err
+	}
+	c.Targets = a.Targets
+	return nil
+}
